Added the current user as an option to send from when sending emails, and changed how sender valeus were read from the form
